% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-method.R
\name{plot}
\alias{plot}
\title{Generic Plot Function for Model Visualizations}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{A model object. Specific methods exist for different model types.}

\item{...}{Additional arguments passed to the plot method.}
}
\value{
A plot object, usually a ggplot, specific to the model type.
}
\description{
\code{plot()} is a generic function designed for visualizing various aspects of
statistical models. It provides a unified interface for generating plots that
vary depending on the model type and the specific method implemented for that
model class.

Current available method:
\itemize{
\item \strong{Confirmatory Factor Analysis}: \code{\link[=plot.lavaan]{plot.lavaan()}} for plotting factor loadings and other metrics for models from the \code{lavaan} package.
}

Future methods may include support for:
\itemize{
\item \strong{Exploratory Factor Analysis}: Models from the \code{psych} package.
\item \strong{Item Response Theory (IRT)}: Models from the \code{mirt} package.
}
}
\keyword{internal}
