% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.compare_model_fit.R
\name{print.compare_model_fit}
\alias{print.compare_model_fit}
\title{Print Method for compare_model_fit Objects}
\usage{
\method{print}{compare_model_fit}(x, digits = 3, p_digits = 3, format = "text", ...)
}
\arguments{
\item{x}{An object of class `compare_model_fit`, created by the `compare_model_fit()` function.}

\item{digits}{Integer. Number of digits to use for displaying numeric values. Defaults to 3.}

\item{p_digits}{Integer. Number of digits to use for displaying p-values. Defaults to 3.}

\item{format}{Character. The format in which to print the table. Options are `"text"`, `"markdown"`, or `"html"`. Defaults to `"text"`.}

\item{...}{Additional arguments passed to `insight::export_table()`.}
}
\value{
Invisibly returns the `compare_model_fit` object. The main purpose of this function is to print the formatted comparison table.
}
\description{
This function provides a custom print method for objects of class
`compare_model_fit`, which are generated by the `compare_model_fit()` function.
It formats and displays the comparison of model fit indices in a clear and
readable table.
}
\examples{
library(psymetrics)
library(lavaan)

model1 <- 'visual  =~ x1 + x2 + x3'
model2 <- 'visual  =~ x1 + x2 + x3 + x4'
fit1 <- cfa(model1, data = HolzingerSwineford1939, estimator = "MLR")
fit2 <- cfa(model2, data = HolzingerSwineford1939, estimator = "MLR")
comparison <- compare_model_fit(fit1, fit2)
print(comparison, digits = 2)
}
