% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_estimates-methods.R
\name{print.model_estimates}
\alias{print.model_estimates}
\title{Print Method for model_estimates Objects}
\usage{
\method{print}{model_estimates}(x, digits = 3, ci_digits = digits, p_digits = 3, ...)
}
\arguments{
\item{x}{An object of class \code{model_estimates}.}

\item{digits}{Number of digits for rounding numeric values.
Default is 3.}

\item{ci_digits}{Number of digits for confidence intervals.
Default is \code{digits}.}

\item{p_digits}{Number of digits for p-values. Default is 3.}

\item{...}{Other arguments are ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
\code{print.model_estimates()} prints a text summary of model
parameter estimates to the console in component blocks.
}
\seealso{
\code{\link[=format_results]{format_results()}}
}
