% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.model_fit.R
\name{print.model_fit}
\alias{print.model_fit}
\title{Print Method for Model Fit Indices}
\usage{
\method{print}{model_fit}(x, digits = 3, p_digits = 3, format = "text", align = "firstleft", ...)
}
\arguments{
\item{x}{An object of class \code{model_fit}, typically created by the \code{\link{model_fit}} function.}

\item{digits}{Number of digits for rounding numeric values. Default is 3.}

\item{p_digits}{Number of digits for rounding p-values. Default is 3.}

\item{format}{Output format for the table. Can be "text", "markdown" (or "md"), or "html". Default is "text".}

\item{align}{Alignment for the table when the format is "text" or "markdown". Options include "left", "right",
"center", and "firstleft". Default is "firstleft". For more information on alignment options, consult
the documentation for \code{\link[insight:export_table]{insight::export_table}}.}

\item{...}{Additional arguments passed to formatting functions.
If \code{format = "text"} or \code{format = "markdown"} they are passed to \code{\link[insight:export_table]{insight::export_table}}.
If \code{format = "html"} these arguments are passed to \code{\link[tinytable:tt]{tinytable::tt}}.}
}
\value{
The formatted table is printed to the console.
}
\description{
\code{print.model_fit} is a custom print method for objects generated by the \code{\link{model_fit}} function.
It formats the output table using \code{\link[insight:format_table]{insight::format_table}} and exports it using \code{\link[insight:export_table]{insight::export_table}}.
}
\examples{
if (requireNamespace("lavaan", quietly = TRUE)) {
  library(lavaan)
  library(psymetrics)

  model <- 'visual  =~ x1 + x2 + x3
          textual =~ x4 + x5 + x6
          speed   =~ x7 + x8 + x9'

  fit <- cfa(model, data = HolzingerSwineford1939,
             estimator = "MLR")
  result <- model_fit(fit)
  result
  print(result, format = "html")
} else {
  message("Please install 'lavaan' to run this example.")
}

}
