% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_model_fit.R
\name{compare_model_fit}
\alias{compare_model_fit}
\title{Compare Model Fit Indices Across Multiple Models}
\usage{
compare_model_fit(..., type = NULL, metrics = "essential", verbose = TRUE)
}
\arguments{
\item{...}{Two or more model objects to be compared.}

\item{type}{A character string specifying the type of fit indices to extract.
Options are \code{"standard"}, \code{"scaled"}, and \code{"robust"}. Defaults to \code{NULL},
which automatically selects \code{"scaled"} if a robust estimator is used, otherwise \code{"standard"}.}

\item{metrics}{A character vector specifying which fit indices to extract.
Defaults to \code{"essential"}, or a custom vector of indices.}

\item{verbose}{Logical. If \code{TRUE}, prints messages about the indices being adjusted.}
}
\value{
A data frame containing the fit indices for each model, with an additional column identifying the models.
}
\description{
\code{compare_model_fit()} compares the fit indices of two or more
models. It extracts the fit indices using \code{\link{model_fit}} and combines them into
a single data frame for easy comparison.
}
\examples{
if (requireNamespace("lavaan", quietly = TRUE)) {
  library(lavaan)
  library(psymetrics)
  model1 <- 'visual  =~ x1 + x2 + x3 + x4'
  model2 <- 'visual  =~ x1 + x2 + x3 + x4 + x5'
  fit1 <- cfa(model1, data = HolzingerSwineford1939, estimator = "MLR")
  fit2 <- cfa(model2, data = HolzingerSwineford1939, estimator = "MLR")
  compare_model_fit(fit1, fit2)
} else {
  message("Please install 'lavaan' to run this example.")
}
}
