% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.compare_model_fit.R
\name{print.compare_model_fit}
\alias{print.compare_model_fit}
\title{Print Method for compare_model_fit Objects}
\usage{
\method{print}{compare_model_fit}(x, digits = 3, p_digits = 3, format = "text", align = "firstleft", ...)
}
\arguments{
\item{x}{An object of class \code{model_fit}, typically created
by the \code{\link{model_fit}} function.}

\item{digits}{Number of digits for rounding numeric values.
Default is 3.}

\item{p_digits}{Number of digits for rounding p-values.
Default is 3.}

\item{format}{Output format for the table. Can be "text",
"markdown" (or "md"), or "html". Default is "text".}

\item{align}{Alignment for the table when the format is
"text" or "markdown". Options include "left", "right",
"center", and "firstleft". Default is "firstleft". For
more information on alignment options, consult the
documentation for \code{\link[insight:export_table]{insight::export_table}}.}

\item{...}{Additional arguments passed to formatting functions.
\itemize{
\item If \code{format = "text"} or \code{format = "markdown"} they are
passed to \code{\link[insight:export_table]{insight::export_table}}.
\item If \code{format = "html"} these arguments are passed to
\code{\link[tinytable:tt]{tinytable::tt}}.
}}
}
\value{
Invisibly returns the \code{compare_model_fit} object.
The main purpose of this function is to print the
formatted comparison table.
}
\description{
This function provides a custom print method for objects
of class \code{compare_model_fit}, which are generated by
the \code{\link[=compare_model_fit]{compare_model_fit()}} function. It formats and
displays the comparison of model fit indices in a clear and
readable table.
}
\examples{
if (requireNamespace("lavaan", quietly = TRUE)) {
  library(psymetrics)
  library(lavaan)

  model1 <- 'visual  =~ x1 + x2 + x3
           textual =~ x4 + x5 + x6'
  fit1 <- cfa(model1, data = HolzingerSwineford1939,
              estimator = "ML")
  fit2 <- cfa(model1, data = HolzingerSwineford1939,
              estimator = "MLR")
  comparison <- compare_model_fit(fit1, fit2)
  comparison
  print(comparison, format = "html")
} else {
  message("Please install 'lavaan' to run this example.")
}
}
