% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.model_fit.R
\name{print.model_fit}
\alias{print.model_fit}
\title{Print Method for Model Fit Indices}
\usage{
\method{print}{model_fit}(x, digits = 3, p_digits = 3, format = "text", ...)
}
\arguments{
\item{x}{An object of class `model_fit`, typically created by the `model_fit` function.}

\item{digits}{Number of digits for rounding numeric values. Default is 3.}

\item{p_digits}{Number of digits for rounding p-values. Default is 3.}

\item{format}{Output format for the table. Can be "text", "markdown" (or "md"), or "html". Default is "text".}

\item{...}{Additional arguments passed to `insight::export_table`.}
}
\value{
The formatted table is printed to the console.
}
\description{
`print.model_fit` is a custom print method for objects generated by the `model_fit` function.
It formats the output table using `insight::format_table` and exports it using `insight::export_table`.
}
\examples{
\dontrun{
library(lavaan)
model <- 'visual  =~ x1 + x2 + x3
          textual =~ x4 + x5 + x6
          speed   =~ x7 + x8 + x9'
fit <- cfa(model, data = HolzingerSwineford1939, estimator = "MLR")
result <- model_fit(fit)
print(result)
}
}
